Reduced coupling and fixed bug
The player's activate_powerup method is now a Kivy event that is
triggered when a cell with a powerup pellet no longer has a pellet,
rather than the character accessing the cell to check if it has a power
up in.
Whether the enemy uses its retreat movement is now determined by the
enemy in its move method rather than the game.
Fixed bug where clicking login buttons multiple times would cause
multiple levels to be generated.